server/requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.18
pydantic>=2.10.0
pydantic-settings>=2.6.1

# Database
motor>=3.5.0
pymongo>=4.8.0

# Computer Vision
opencv-python>=4.10.0
ultralytics>=8.3.30
torch>=2.5.0
torchvision>=0.20.0
pillow>=11.0.0

# OCR
easyocr>=1.7.2

# AI/ML
google-generativeai>=0.8.3

# Roboflow
roboflow>=1.1.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
